Add more access checks (#53)

+++ apps/api/src/resolvers/deleteCheckpoint.js
@@ -1,4 +1,28 @@
-module.exports = async (_, { id }, { dataSources: { db }}) => {
+module.exports = async (_, { id }, { user, dataSources: { db }}) => {
+  const checkpoint = await db.findCheckpointById(id)
+
+  if (!checkpoint) {
+    return {
+      success: false
+    }
+  }
+
+  const event = await db.findEventById(checkpoint.event_id)
+
+  if (!event) {
+    return {
+      success: false
+    }
+  }
+
+  const participant = await db.findParticipant({ userId: user.id, eventId: event.id })
+
+  if (!participant) {
+    return {
+      success: false
+    }
+  }
+
   await db.deleteCheckpoint(id)
 
   return {
